# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SENTINEL",
    "description": "Real-time application fraud detection for digital lending.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/decisions/application": {
      "post": {
        "tags": [
          "decisions"
        ],
        "summary": "Score a credit application and persist the decision",
        "operationId": "decide_application_api_v1_decisions_application_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/decisions/{decision_id}": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "Get Decision",
        "operationId": "get_decision_api_v1_decisions__decision_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decisions/{decision_id}/adverse-action": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "Reg B adverse action notice (declined applications only)",
        "operationId": "get_adverse_action_api_v1_decisions__decision_id__adverse_action_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Decision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdverseActionNotice"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Not a declined application",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decisions": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "List Decisions",
        "operationId": "list_decisions_api_v1_decisions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "band",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Band"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Band"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stream": {
      "get": {
        "tags": [
          "stream"
        ],
        "summary": "Live decision feed (SSE)",
        "operationId": "stream_decisions_api_v1_stream_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT; EventSource cannot send an Authorization header",
              "title": "Token"
            },
            "description": "JWT; EventSource cannot send an Authorization header"
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events. `event: decision` carries a DecisionResponse JSON in `data`; `event: drift` carries a DriftEventOut JSON; `: ping` comments every 15 s keep the connection open.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stream/status": {
      "get": {
        "tags": [
          "stream"
        ],
        "summary": "Stream Status",
        "operationId": "stream_status_api_v1_stream_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamStatus"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/stream/switch": {
      "post": {
        "tags": [
          "stream"
        ],
        "summary": "Switch replay source (admin)",
        "operationId": "switch_source_api_v1_stream_switch_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "base",
                "shift"
              ],
              "type": "string",
              "description": "base = test months of Base; shift = variant file",
              "title": "Source"
            },
            "description": "base = test months of Base; shift = variant file"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamStatus"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stream/start": {
      "post": {
        "tags": [
          "stream"
        ],
        "summary": "Start replay (admin)",
        "operationId": "start_stream_api_v1_stream_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamStatus"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/stream/stop": {
      "post": {
        "tags": [
          "stream"
        ],
        "summary": "Stop replay (admin)",
        "operationId": "stop_stream_api_v1_stream_stop_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamStatus"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/cases/{decision_id}/similar": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Nearest past cases by narrative embedding",
        "operationId": "similar_cases_api_v1_cases__decision_id__similar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Decision Id"
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarCasesResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "501": {
            "description": "Not implemented yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/copilot/ask": {
      "post": {
        "tags": [
          "copilot"
        ],
        "summary": "Ask the read-only analyst copilot about a decision",
        "operationId": "copilot_ask_api_v1_copilot_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotAskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApplicationEvent": {
        "properties": {
          "income": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Income"
          },
          "name_email_similarity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Name Email Similarity"
          },
          "prev_address_months_count": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": -1.0,
            "title": "Prev Address Months Count"
          },
          "current_address_months_count": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": -1.0,
            "title": "Current Address Months Count"
          },
          "customer_age": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 10.0,
            "title": "Customer Age"
          },
          "days_since_request": {
            "type": "number",
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Days Since Request"
          },
          "intended_balcon_amount": {
            "type": "number",
            "maximum": 250.0,
            "minimum": -100.0,
            "title": "Intended Balcon Amount"
          },
          "payment_type": {
            "type": "string",
            "enum": [
              "AA",
              "AB",
              "AC",
              "AD",
              "AE"
            ],
            "title": "Payment Type"
          },
          "zip_count_4w": {
            "type": "integer",
            "maximum": 20000.0,
            "minimum": 0.0,
            "title": "Zip Count 4W"
          },
          "velocity_6h": {
            "type": "number",
            "maximum": 50000.0,
            "minimum": -1000.0,
            "title": "Velocity 6H"
          },
          "velocity_24h": {
            "type": "number",
            "maximum": 50000.0,
            "minimum": 0.0,
            "title": "Velocity 24H"
          },
          "velocity_4w": {
            "type": "number",
            "maximum": 50000.0,
            "minimum": 0.0,
            "title": "Velocity 4W"
          },
          "bank_branch_count_8w": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Bank Branch Count 8W"
          },
          "date_of_birth_distinct_emails_4w": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Date Of Birth Distinct Emails 4W"
          },
          "employment_status": {
            "type": "string",
            "enum": [
              "CA",
              "CB",
              "CC",
              "CD",
              "CE",
              "CF",
              "CG"
            ],
            "title": "Employment Status"
          },
          "credit_risk_score": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": -500.0,
            "title": "Credit Risk Score"
          },
          "email_is_free": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Email Is Free"
          },
          "housing_status": {
            "type": "string",
            "enum": [
              "BA",
              "BB",
              "BC",
              "BD",
              "BE",
              "BF",
              "BG"
            ],
            "title": "Housing Status"
          },
          "phone_home_valid": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Phone Home Valid"
          },
          "phone_mobile_valid": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Phone Mobile Valid"
          },
          "bank_months_count": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": -1.0,
            "title": "Bank Months Count"
          },
          "has_other_cards": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Has Other Cards"
          },
          "proposed_credit_limit": {
            "type": "number",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Proposed Credit Limit"
          },
          "foreign_request": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Foreign Request"
          },
          "source": {
            "type": "string",
            "enum": [
              "INTERNET",
              "TELEAPP"
            ],
            "title": "Source"
          },
          "session_length_in_minutes": {
            "type": "number",
            "maximum": 500.0,
            "minimum": -1.0,
            "title": "Session Length In Minutes"
          },
          "device_os": {
            "type": "string",
            "enum": [
              "windows",
              "macintosh",
              "linux",
              "x11",
              "other"
            ],
            "title": "Device Os"
          },
          "keep_alive_session": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Keep Alive Session"
          },
          "device_distinct_emails_8w": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": -1.0,
            "title": "Device Distinct Emails 8W"
          },
          "device_fraud_count": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Device Fraud Count"
          },
          "month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 7.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "device_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Id"
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 1,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "pattern": "^\\+?[0-9 ()\\-]{7,31}$",
            "title": "Phone"
          },
          "ip": {
            "type": "string",
            "maxLength": 45,
            "minLength": 1,
            "title": "Ip"
          },
          "address": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Address"
          },
          "applicant_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Applicant Name"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9._:-]{1,64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "history": {
            "type": "boolean",
            "title": "History",
            "default": false
          },
          "fraud_bool": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  0,
                  1
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Fraud Bool"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "income",
          "name_email_similarity",
          "prev_address_months_count",
          "current_address_months_count",
          "customer_age",
          "days_since_request",
          "intended_balcon_amount",
          "payment_type",
          "zip_count_4w",
          "velocity_6h",
          "velocity_24h",
          "velocity_4w",
          "bank_branch_count_8w",
          "date_of_birth_distinct_emails_4w",
          "employment_status",
          "credit_risk_score",
          "email_is_free",
          "housing_status",
          "phone_home_valid",
          "phone_mobile_valid",
          "bank_months_count",
          "has_other_cards",
          "proposed_credit_limit",
          "foreign_request",
          "source",
          "session_length_in_minutes",
          "device_os",
          "keep_alive_session",
          "device_distinct_emails_8w",
          "device_fraud_count",
          "device_id",
          "email",
          "phone",
          "ip",
          "address",
          "applicant_name"
        ],
        "title": "ApplicationEvent",
        "description": "A raw credit application: every BAF column plus synthetic identifiers used for the entity graph.\n\nProtected attributes (customer_age, employment_status, income) are accepted for fairness evaluation\nbut are never model features (enforced in ml/featurize.py)."
      },
      "Band": {
        "type": "string",
        "enum": [
          "APPROVE",
          "STEP_UP",
          "REVIEW",
          "DECLINE"
        ],
        "title": "Band"
      },
...
      "DecisionResponse": {
        "properties": {
          "decision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Decision Id"
          },
          "application_id": {
            "type": "string",
            "format": "uuid",
            "title": "Application Id"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "score": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "band": {
            "$ref": "#/components/schemas/Band"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionOutcome"
          },
          "reason_codes": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "type": "array",
            "title": "Reason Codes"
          },
          "graph_signals": {
            "$ref": "#/components/schemas/GraphSignals"
          },
          "graph_uplift": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Graph Uplift"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "decision_id",
          "application_id",
          "score",
          "band",
          "decision",
          "reason_codes",
          "graph_signals",
          "graph_uplift",
          "model_version",
          "latency_ms",
          "created_at"
        ],
        "title": "DecisionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT from POST /api/v1/auth/login",
        "scheme": "bearer"
      }
    }
  }
}